Placed modal-backdrop in between Coroprate Header and Modal.
As the mobile corporate-header is on z-index is on layer 1100, both modal and modal-backdrop has to be placed on top of them. As of now, the backdrop is behind the corporate-header and the modal on top.

Would really see the corporate-header move down the z-index scale and keep the original Bootstrap modal z values. Not sure what will happen to z-indexed stuff in applications.  But we'll find out with 3.4.0 as the modal already now is at 1200.